% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_ll}
\alias{calc_ll}
\title{function to calculate ll}
\usage{
calc_ll(di, l, r, t, pop_size, freq_ancestor_1, condition)
}
\arguments{
\item{di}{distance}

\item{l}{left}

\item{r}{right}

\item{t}{time}

\item{pop_size}{population size}

\item{freq_ancestor_1}{freq ancestor}

\item{condition}{conditioning}
}
\description{
function to calculate ll
}
