% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count_of_states_unphased}
\alias{count_of_states_unphased}
\title{create a matrix with the frequency of each state}
\usage{
count_of_states_unphased(local_anc_matrix)
}
\arguments{
\item{local_anc_matrix}{local ancestry matrix}
}
\description{
create a matrix with the frequency of each state
}
