% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_time_markers.R
\name{estimate_time_markers}
\alias{estimate_time_markers}
\title{Estimate the time since the onset of hybridization, using the observed number of junctions, taking into account the distribution of markers}
\usage{
estimate_time_markers(J = NA, N = Inf, H_0 = 0.5,
  marker_distribution = NA)
}
\arguments{
\item{J}{The observed number of junctions}

\item{N}{Population Size}

\item{H_0}{Frequency of heterozygosity at t = 0}

\item{marker_distribution}{A vector containing the position of all markers in Morgan.}
}
\value{
The number of generations passed since the onset of hybridization
}
\description{
Estimate the time since the onset of hybridization, following equation 1 in Janzen et al. unpublished
}
\examples{
markers <- seq(from = 0, to = 1, length.out = 100)
J <- number_of_junctions_markers(N = 100, H_0 = 0.5, t = 200, marker_distribution = markers)
estimate_time_markers(J = J,
                     N = 100,
H_0 = 0.5,
marker_distribution = markers) #should be 200 again
}
