% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_prob_from_matrix_cpp}
\alias{get_prob_from_matrix_cpp}
\title{function to calculate prob}
\usage{
get_prob_from_matrix_cpp(left, right, p, P)
}
\arguments{
\item{p}{freq}

\item{P}{seven states output}

\item{l}{left}

\item{r}{right}
}
\description{
function to calculate prob
}
