\name{junctions-package}
\alias{junctions-package}
\alias{junctions}
\docType{package}
\title{
  Extending The Theory of Junctions 
}
\description{
  The theory of junctions is extended by this package by including the effect of a finite number of recombination sites along the chromosome. The package provides functions to calculate the estimated number of junctions, depending on the time since the onset of hybridization, population size, number of recombination sites, initial heterozygosity and the number of crossovers per meiosis. 
}
\details{
  TOWRITE
}
\author{
Maintainer: Thijs Janzen <thijsjanzen@gmail.com>
}
\references{
  Janzen, Thijs, Arne W. Nolte, and Arne Traulsen. "Haplotype block dynamics in hybrid populations." bioRxiv (2016): 058107.
}
\keyword{ junctions }

