% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikelihood_unphased_cpp}
\alias{loglikelihood_unphased_cpp}
\title{function to calculate log likelihood using cpp}
\usage{
loglikelihood_unphased_cpp(
  local_anc_matrix,
  locations,
  pop_size,
  freq_ancestor_1,
  t,
  phased,
  verbose = FALSE,
  num_threads = 1L
)
}
\arguments{
\item{local_anc_matrix}{local ancestry matrix}

\item{locations}{locations of markers}

\item{pop_size}{population size}

\item{freq_ancestor_1}{frequency of the most common ancestor}

\item{t}{time}

\item{phased}{is the data phased or not?}

\item{verbose}{verbose output}

\item{num_threads}{number of threads, default is one thread. Set to -1 to
use all available threads.}
}
\description{
function to calculate log likelihood using cpp
}
