% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{matrix_of_states_unphased}
\alias{matrix_of_states_unphased}
\title{create a matrix with the total contribution of each term to the likelihood}
\usage{
matrix_of_states_unphased(
  local_anc_matrix,
  locations,
  pop_size,
  freq_ancestor_1 = 0.5,
  t
)
}
\arguments{
\item{local_anc_matrix}{local ancestry matrix}

\item{locations}{locations in Morgan}

\item{pop_size}{population size}

\item{freq_ancestor_1}{frequency of ancestor 1}

\item{t}{time point}
}
\description{
create a matrix with the total contribution of each term to the likelihood
}
