% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_phased_unphased_cpp}
\alias{sim_phased_unphased_cpp}
\title{simulate junctions}
\usage{
sim_phased_unphased_cpp(pop_size, freq_ancestor_1, total_runtime,
  size_in_morgan, number_of_markers, time_points, seed, verbose)
}
\arguments{
\item{pop_size}{Population Size}

\item{freq_ancestor_1}{Frequency of ancestor 1 at t = 0}

\item{total_runtime}{Maximum time after which the simulation is to be stopped}

\item{size_in_morgan}{Mean number of crossovers per meiosis (e.g. size in Morgan of the chromosome)}

\item{number_of_markers}{The number of genetic markers superimposed on the chromosome.}

\item{time_points}{vector with time points at which local ancestry has to be recorded to be returned at the end of the simulation. If left at -1, ancestry is recorded at every generation (computationally heavy).}

\item{seed}{Seed of the pseudo-random number generator}

\item{verbose}{displays a progress bar}
}
\description{
simulate junctions
}
