% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_state_cpp}
\alias{single_state_cpp}
\title{function to calculate 7 states}
\usage{
single_state_cpp(t, N, d)
}
\arguments{
\item{t}{time}

\item{N}{pop size}

\item{d}{distance}
}
\description{
function to calculate 7 states
}
