% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unphased_log_likelihood.R
\name{unphased_log_likelihood}
\alias{unphased_log_likelihood}
\title{calculate the log likelihood of observing the data}
\usage{
unphased_log_likelihood(local_anc, locations, pop_size,
  freq_ancestor_1 = 0.5, t)
}
\arguments{
\item{local_anc}{matrix with two columns, where the first column represents ancestry on chromosome 1, and the second column represents ancestry on chromosome 2. Ancestry labels used should be [0, 1], where 0 indicates the first ancestor, and 1 indicates the second ancestor.}

\item{locations}{locations of the used markers (in Morgan)}

\item{pop_size}{population size}

\item{freq_ancestor_1}{Frequency of ancestor 1 at t = 0}

\item{t}{time since admixture}
}
\value{
log likelihood
}
\description{
Calculates the log likelihood of observing the unphased data, given the population size, initial heterozygosity and time since admixture
}
