% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_time_unphased.R
\name{estimate_time_unphased}
\alias{estimate_time_unphased}
\title{estimates the time since admixture, given unphased ancestry data.}
\usage{
estimate_time_unphased(local_anc, locations, pop_size, freq_ancestor_1,
  max_t, optim_pop_size = FALSE, verbose = FALSE)
}
\arguments{
\item{local_anc}{Local_anc can be provided as either matrix with two columns, where the first column represents ancestry on chromosome 1, and the second column represents ancestry on chromosome 2. Ancestry labels used should be [0, 1], where 0 indicates the first ancestor, and 1 indicates the second ancestor. Alternatively, the user can provide a vector indicating whether at the specific marker, the focal individual is homozygous for the first ancestor (0), homozygous for the second ancestor (1) or heterozygous (2).}

\item{locations}{locations of the used markers (in Morgan)}

\item{pop_size}{population size}

\item{freq_ancestor_1}{Frequency of ancestor 1 at t = 0}

\item{max_t}{maximum time to be considered for maximum likelihood optimization. Too large values might cause the maximum likelihood algorithm to deviate from the global maximum.}

\item{optim_pop_size}{If TRUE, population size is also optimized. Starting point of the optimizaton will then be on the given population size, and half the maximum time.}

\item{verbose}{display intermediate output? Default = FALSE}
}
\description{
Calculates the time since admixture, given unphased ancestry data.
}
